backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.0.1
channels==4.0.0
celery==5.3.4
django-celery-beat==2.5.0
daphne==4.0.0